My-app-V1/backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-sqlalchemy
opentelemetry-propagator-aws-xray
opentelemetry-sdk-extension-aws
